+++ examples/examples.py
@@ -16,20 +16,21 @@ csv_path = ""
 os.makedirs("./examples/local_db", exist_ok=True)
 
 availabe_examples = {
+    # small examples for easier understanding of different features and configurations
     "small": {"scenario": "example_01a", "study_case": "base"},
     "small_dam": {"scenario": "example_01a", "study_case": "dam"},
     "small_with_opt_clearing": {
         "scenario": "example_01a",
         "study_case": "dam_with_complex_clearing",
     },
-    "small_with_BB_and_LB": {
-        "scenario": "example_01c",
-        "study_case": "dam_with_complex_opt_clearing",
-    },
     "small_with_vre": {"scenario": "example_01b", "study_case": "base"},
     "small_with_vre_and_storage": {
         "scenario": "example_01c",
         "study_case": "eom_only",
+    },
+    "small_with_BB_and_LB": {
+        "scenario": "example_01c",
+        "study_case": "dam_with_complex_opt_clearing",
     },
     "small_with_vre_and_storage_and_complex_clearing": {
         "scenario": "example_01c",
@@ -61,20 +62,19 @@ availabe_examples = {
     },
     "large_2019_eom": {"scenario": "example_02", "study_case": "base_case_2019"},
     "large_2019_eom_crm": {
-        "scenario": "example_02",
+        "scenario": "example_03",
         "study_case": "eom_crm_case_2019",
     },
     "large_2019_day_ahead": {
-        "scenario": "example_02",
+        "scenario": "example_03",
         "study_case": "dam_case_2019",
     },
-    "small_learning_1": {"scenario": "example_02a", "study_case": "base"},
-    "small_learning_2": {"scenario": "example_02b", "study_case": "base"},
-    "small_learning_3": {"scenario": "example_02c", "study_case": "base"},
-    "learning_with_complex_bids": {
-        "scenario": "example_02e",
-        "study_case": "base",
+    "large_2019_with_DSM": {
+        "scenario": "example_03",
+        "study_case": "base_case_2019_with_DSM",
     },
+    "large_2019_rl": {"scenario": "example_03a", "study_case": "base_case_2019"},
+    "large_2021_rl": {"scenario": "example_03b", "study_case": "base_case_2021"},
 }
 
 # %%
@@ -85,6 +85,7 @@ if __name__ == "__main__":
     - timescale: with database and grafana (note: you need docker installed)
     """
     data_format = "local_db"  # "local_db" or "timescale"
+    example = "small"
     example = "small"
 
     if data_format == "local_db":